solution

+++ wallis.c
@@ -27,7 +27,7 @@ float wallis_pi(int n)
   { float p=1;
   for(int i=1;i<=n ;i++)
   {float e=4*i*i;
-  float k=e/(e-1);
+  float k=(float)e/(e-1);
   p=p*k;
   }return p*2;
